sanic/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiofiles==0.6.0
    # via sanic
aiosqlite==0.16.0
    # via tortoise-orm
asyncpg==0.21.0
    # via tortoise-orm
certifi==2020.12.5
    # via httpx
h11==0.9.0
    # via httpcore
httpcore==0.11.1
    # via httpx
httptools==0.1.1
    # via sanic
httpx==0.15.4
    # via sanic
idna==3.1
    # via rfc3986
iso3166==1.0.1
    # via -r requirements.in
iso8601==0.1.13
    # via tortoise-orm
marshmallow-enum==1.4.1
    # via -r requirements.in
marshmallow==2.20.0
    # via
    #   -r requirements.in
    #   marshmallow-enum
multidict==5.1.0
    # via sanic
pypika==0.44.1
    # via tortoise-orm
pytz==2020.5
    # via tortoise-orm
rfc3986[idna2008]==1.4.0
    # via httpx
sanic==20.12.1
    # via -r requirements.in
six==1.15.0
    # via sqlalchemy-utils
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
sqlalchemy-utils==0.36.8
    # via -r requirements.in
sqlalchemy==1.3.22
    # via sqlalchemy-utils
tortoise-orm[asyncpg]==0.16.19
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via aiosqlite
ujson==4.0.1
    # via sanic
uvloop==0.14.0
    # via sanic
websockets==8.1
    # via sanic
